// packages/contracts/typechain-types/CapabilityRegistry.ts
export declare namespace CapabilityRegistry {
  export type ReputationRecordStruct = {
    score: BigNumberish;
    totalJobs: BigNumberish;
    successfulJobs: BigNumberish;
    weightedSuccesses: BigNumberish;
    weightedTotal: BigNumberish;
  };

  export type ReputationRecordStructOutput = [
    score: bigint,
    totalJobs: bigint,
    successfulJobs: bigint,
    weightedSuccesses: bigint,
    weightedTotal: bigint
  ] & {
    score: bigint;
    totalJobs: bigint;
    successfulJobs: bigint;
    weightedSuccesses: bigint;
    weightedTotal: bigint;
  };

  export type AgentProfileStruct = {
    agentAddress: AddressLike;
    name: string;
    capabilities: string[];
    pricePerJob: BigNumberish;
    endpoint: string;
    available: boolean;
    registeredAt: BigNumberish;
    reputation: CapabilityRegistry.ReputationRecordStruct;
  };

  export type AgentProfileStructOutput = [
    agentAddress: string,
    name: string,
    capabilities: string[],
    pricePerJob: bigint,
    endpoint: string,
    available: boolean,
    registeredAt: bigint,
    reputation: CapabilityRegistry.ReputationRecordStructOutput
  ] & {
    agentAddress: string;
    name: string;
    capabilities: string[];
    pricePerJob: bigint;
    endpoint: string;
    available: boolean;
    registeredAt: bigint;
    reputation: CapabilityRegistry.ReputationRecordStructOutput;
  };
}

export interface CapabilityRegistryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "agentList"
      | "getProfile"
      | "jobEscrow"
      | "lookup"
      | "owner"
      | "register"
      | "setAvailable"
      | "setJobEscrow"
      | "updateReputation"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "AgentRegistered" | "ReputationUpdated"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "agentList",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getProfile",
    values: [AddressLike]
  ): string;
  encodeFunctionData(functionFragment: "jobEscrow", values?: undefined): string;
  encodeFunctionData(functionFragment: "lookup", values: [string]): string;
  encodeFunctionData(functionFragment: "owner", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "register",
    values: [string, string[], BigNumberish, string]
  ): string;
  encodeFunctionData(
    functionFragment: "setAvailable",
    values: [boolean]
  ): string;
  encodeFunctionData(
    functionFragment: "setJobEscrow",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "updateReputation",
    values: [AddressLike, boolean, BigNumberish]
  ): string;

  decodeFunctionResult(functionFragment: "agentList", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getProfile", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "jobEscrow", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "lookup", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "owner", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "register", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "setAvailable",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "setJobEscrow",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "updateReputation",
    data: BytesLike
  ): Result;
}

export namespace AgentRegisteredEvent {
  export type InputTuple = [
    agent: AddressLike,
    name: string,
    capabilities: string[]
  ];
  export type OutputTuple = [
    agent: string,
    name: string,
    capabilities: string[]
  ];
  export interface OutputObject {
    agent: string;
    name: string;
    capabilities: string[];
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace ReputationUpdatedEvent {
  export type InputTuple = [
    agent: AddressLike,
    newScore: BigNumberish,
    success: boolean,
    jobValue: BigNumberish
  ];
  export type OutputTuple = [
    agent: string,
    newScore: bigint,
    success: boolean,
    jobValue: bigint
  ];
  export interface OutputObject {
    agent: string;
    newScore: bigint;
    success: boolean;
    jobValue: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface CapabilityRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): CapabilityRegistry;
  waitForDeployment(): Promise<this>;

  interface: CapabilityRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  agentList: TypedContractMethod<[arg0: BigNumberish], [string], "view">;

  getProfile: TypedContractMethod<
    [agent: AddressLike],
    [CapabilityRegistry.AgentProfileStructOutput],
    "view"
  >;

  jobEscrow: TypedContractMethod<[], [string], "view">;

  lookup: TypedContractMethod<
    [capability: string],
    [CapabilityRegistry.AgentProfileStructOutput[]],
    "view"
  >;

  owner: TypedContractMethod<[], [string], "view">;

  register: TypedContractMethod<
    [
      name: string,
      capabilities: string[],
      pricePerJob: BigNumberish,
      endpoint: string
    ],
    [void],
    "nonpayable"
  >;

  setAvailable: TypedContractMethod<[available: boolean], [void], "nonpayable">;

  setJobEscrow: TypedContractMethod<
    [_jobEscrow: AddressLike],
    [void],
    "nonpayable"
  >;

  updateReputation: TypedContractMethod<
    [agent: AddressLike, success: boolean, jobValue: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "agentList"
  ): TypedContractMethod<[arg0: BigNumberish], [string], "view">;
  getFunction(
    nameOrSignature: "getProfile"
  ): TypedContractMethod<
    [agent: AddressLike],
    [CapabilityRegistry.AgentProfileStructOutput],
    "view"
  >;
  getFunction(
    nameOrSignature: "jobEscrow"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "lookup"
  ): TypedContractMethod<
    [capability: string],
    [CapabilityRegistry.AgentProfileStructOutput[]],
    "view"
  >;
  getFunction(
    nameOrSignature: "owner"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "register"
  ): TypedContractMethod<
    [
      name: string,
      capabilities: string[],
      pricePerJob: BigNumberish,
      endpoint: string
    ],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "setAvailable"
  ): TypedContractMethod<[available: boolean], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "setJobEscrow"
  ): TypedContractMethod<[_jobEscrow: AddressLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "updateReputation"
  ): TypedContractMethod<
    [agent: AddressLike, success: boolean, jobValue: BigNumberish],
    [void],
    "nonpayable"
  >;

  getEvent(
    key: "AgentRegistered"
  ): TypedContractEvent<
    AgentRegisteredEvent.InputTuple,
    AgentRegisteredEvent.OutputTuple,
    AgentRegisteredEvent.OutputObject
  >;
  getEvent(
    key: "ReputationUpdated"
  ): TypedContractEvent<
    ReputationUpdatedEvent.InputTuple,
    ReputationUpdatedEvent.OutputTuple,
    ReputationUpdatedEvent.OutputObject
  >;

  filters: {
    "AgentRegistered(address,string,string[])": TypedContractEvent<
      AgentRegisteredEvent.InputTuple,
      AgentRegisteredEvent.OutputTuple,
      AgentRegisteredEvent.OutputObject
    >;
    AgentRegistered: TypedContractEvent<
      AgentRegisteredEvent.InputTuple,
      AgentRegisteredEvent.OutputTuple,
      AgentRegisteredEvent.OutputObject
    >;

    "ReputationUpdated(address,uint256,bool,uint256)": TypedContractEvent<
      ReputationUpdatedEvent.InputTuple,
      ReputationUpdatedEvent.OutputTuple,
      ReputationUpdatedEvent.OutputObject
    >;
    ReputationUpdated: TypedContractEvent<
      ReputationUpdatedEvent.InputTuple,
      ReputationUpdatedEvent.OutputTuple,
      ReputationUpdatedEvent.OutputObject
    >;
  };
}